Give Xcode more permissions in the sandbox.
Grant access to Xcode's caches, as well as the `worksrcpath`. If the
ports DB or the build directory* are on other volumes, grant access to
the `TemporaryItems` directory of that volume as well. Grant permission
to create jobs.

\* Since I have 96 GiB of RAM, I have placed the build directory on a
tmpfs to avoid wear and tear on my SSD. I just hope that no port needs
more than 48 GiB to build...